disable attached objects filter
This is necessary because the attached objects filter removes the color
from the point cloud, and tabletop_object_detector can only handle
colored point clouds. With this commit, we can remove our patched
tabletop_object_detector overlay. On the down side, we'll probably have
to pay attention that the object in the gripper is not visible in the
Kinect data when planning an arm movement (otherwise we'll get "starting
state is in collision").